--- a/.dockerignore
+++ b/.dockerignore
@@ -4,3 +4,4 @@
 /.git
 /dist
 .env
+.var