This PR unblocks deployments.

1. Vendor in the JSON database implementation from Chainsauce, as it is not a part of the public interface anymore, it should also make things easier for us
2. Fix a big regression in performance introduced in https://github.com/gitcoinco/grants-stack-indexer/pull/315 that only surfaced in Fly with the non-blocking I/O version of the JSON database
  - The delayed writes are unbounded and they seem to be saturating the I/O in Fly, introducing a queue with a maximum concurrency of 5 works much better for Fly and doesn't affect the indexing time in my machine
  - Revert the mkdir call to happen with the delayed write instead of when reading the file, it is an expensive call for the contributions dataset, which is 6 directories deep and most contributions are new, so we're mostly creating directories and new files
3. Introduced a convenient `flushWrites` function to the database that can be awaited

Snapshots are good against `release`

----

Notes:

I created an instance in indexer-staging with an unencrypted volume to see if performance was better, it didn't seem to affect the index time and this is what I get with a simple write speed test:

```bash
# unecrypted volume
root@32874e1ef114e8:/usr/src/app# dd if=/dev/zero of=/mnt/indexer/test bs=1G count=1 oflag=dsync
1+0 records in
1+0 records out
1073741824 bytes (1.1 GB, 1.0 GiB) copied, 3.53414 s, 304 MB/s

# encrypted volume
root@3287eeeb342085:/usr/src/app#  dd if=/dev/zero of=/mnt/indexer/test bs=1G count=1 oflag=dsync
1+0 records in
1+0 records out
1073741824 bytes (1.1 GB, 1.0 GiB) copied, 2.29078 s, 469 MB/s
```

Speeds vary between 300 MB/s to 700 MB/s for both instances, I got 140 MB/s in a Google Cloud instance I had at hand, so the write speed seems reasonable.

I think it will be much clearer once we try a proper database, I know the JSON storage is not the most efficient but it's surprising to me that indexing in my laptop can take minutes and take up to 40 minutes in Fly, but it makes me wonder if this is as far as we'll get in Fly and whether this is another reason to look for another provider, I know we've already had a lot of frustrations with it but for me the biggest one is that all the work to speed up indexing seems to not carry over when running in Fly, we're on the performance CPU class and there are no other disk options.

I had a quick look at Railway but their [persistent storage offering is very lacking](https://docs.railway.app/reference/volumes#caveats), same thing [with Render](https://render.com/docs/disks#application-considerations), these two are the main competitors to Fly.

I will open a meta issue so we can discuss the situation with Fly, the switch to a database + resumable indexing will reduce these issues but it's worth having a conversation, another way to alleviate these issues would be to just use a server database like Postgres, but that comes with it's own set of trade offs, and I am not sure I like the idea of having the platform dictate our design choices, a disk based database should be perfectly reasonable.


----

```
flyctl -c fly.staging.toml deploy --wait-timeout=3600  43.56s user 22.16s system 1% cpu 1:08:52.47 total
```


---

**Conclusion**

I tested indexing in a DO droplet with their most performant offering with a dedicated CPU and got similar results as Fly, my conclusion is that the best thing to do is to continue with the work on the database and leave the JSON storage as is for now as we're just getting diminishing returns with these optimizations, but this PR is necessary to unblock while that's being done: https://github.com/gitcoinco/grants-stack-indexer/pull/338